moved details to specific files
