+++ webofneeds/won-owner-webapp/src/main/webapp/webpack.config.ts
@@ -55,6 +55,24 @@ function config(env, argv): Configuration {
           "config",
           `usecase-definitions.js`
         ),
+        personDetails$: path.resolve(
+          __dirname,
+          "config",
+          "details",
+          `person.js`
+        ),
+        musicianDetails$: path.resolve(
+          __dirname,
+          "config",
+          "details",
+          `musician.js`
+        ),
+        realEstateDetails$: path.resolve(
+          __dirname,
+          "config",
+          "details",
+          `real-estate.js`
+        ),
         config$: path.resolve(__dirname, "config", `${nodeEnv}.js`),
         jsonld$: require.resolve("jsonld/dist/jsonld.js"), // This is needed because `jsonld`s entrypoint is not compiled to compatible js (uses spread operators). With this resolve hook we instead use the compiled version.
       },
